Add recipe-based inventory (BOM), worker bottom sheet, and salary validation
- Recipe/BOM system: ingredient-level stock deduction with unit conversions
- Stock quantities: Int → Double throughout backend and client
- Per-item stockBehavior: NONE/DIRECT/RECIPE for flexible stock tracking
- New RecipesTable, RecipeIngredientsTable with full CRUD API
- Order stock enforcement branches on stockBehavior (DIRECT/RECIPE/NONE)
- Worker dialog: AlertDialog → ModalBottomSheet with modern UI
- Worker form data persists on dismiss (reopen without data loss)
- Salary amount now mandatory (> 0) for worker creation and editing
- String resources added in EN and AR for both stock and staff features

+++ core/core-network/src/commonMain/kotlin/net/marllex/waselak/core/network/dto/AnalyticsDashboardDto.kt
@@ -205,8 +205,8 @@ data class AlertsResponse(
 data class StockOverviewItemResponse(
     @SerialName("stock_id") val stockId: String,
     @SerialName("item_name") val itemName: String,
-    val quantity: Int,
-    @SerialName("min_quantity") val minQuantity: Int,
+    val quantity: Double,
+    @SerialName("min_quantity") val minQuantity: Double,
     @SerialName("cost_price") val costPrice: Double,
     val unit: String,
     val status: String,
@@ -215,8 +215,8 @@ data class StockOverviewItemResponse(
 @Serializable
 data class StockMovementResponse(
     val date: String,
-    val added: Int,
-    val deducted: Int,
+    val added: Double,
+    val deducted: Double,
 )
 
 @Serializable
